rename camel case file name and remove old code
