+++ lib/models/game_master.js
@@ -4,8 +4,8 @@ const EventEmitter2 = require('eventemitter2');
 const Game = require('./game');
 const Player = require('./player');
 const Map = require('./map');
-const NormalItem = require('./normalItem');
-const PowerItem = require('./powerItem');
+const NormalItem = require('./normal_item');
+const PowerItem = require('./power_item');
 const EnemyAI1 = require('./enemy_ai1');
 
 const PLAYER_SIZE = 5;
