+++ lib/models/power_item.js
