+++ views/index.ejs
@@ -8,20 +8,6 @@
     <title>パクパクモグモグ</title>
 </head>
 <body>
-<!--
-<div id='inputDiv'>
-<div id='usernameDiv'>
-Username : <input type=text id='username' value='input your name' />
-</div>
-<br>
-<button type="button" id="enter">Enter</button>
-<button type="button" id='start' style='display:none'>Start</button>
-<button type="button" id='back' style='display:none'>Back</button>
-<div id='userList' style='display:none'></div>
-<img id="wolfImg" src="/img/wolf.png" >
-<img id="sheepImg" src="/img/sheep.png" >
-</div>
--->
 <script src="/socket.io/socket.io.js"></script>
 <script src="/vendor/enchant.js"></script>
 <script src="/js/socket.js"></script>
